admin page to approve stores modified to incorporate gift and unknown cases

+++ app/controllers/admin/stores_controller.rb
@@ -20,22 +20,36 @@ class Admin::StoresController < ApplicationController
 
   # Update the store name. If the updated name is already in the
   # database, then delete the store and move all its products to
-  # the existing store
+  # the existing store. Also handle cases for unknowns and gifts
   #
   def update
     store           = Store.find(params[:id])
     fetched_store   = Store.fetch(params[:name])
-    
-    if fetched_store.nil? || store.id == fetched_store.id
-      params[:is_approved] = true
-      store.edit(params)
+    type            = params[:name].downcase.to_sym
 
-      @store = store
-    else
-      store.move_products_to(fetched_store)
+    case type
+    when :gift
+      store.change_products_to_gift
       store.destroy
 
-      @store = fetched_store
+      @store = nil
+    when :unknown
+      store.change_products_store_to_unknown
+      store.destroy
+
+      @store = nil
+    else
+      if fetched_store.nil? || store.id == fetched_store.id
+        params[:is_approved] = true
+        store.edit(params)
+
+        @store = store
+      else
+        store.move_products_to(fetched_store)
+        store.destroy
+
+        @store = fetched_store
+      end
     end
   rescue => ex
     handle_exception(ex)
